+++ packages/client/src/components/report/ReportDetails.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 
 import Stack from '../layout-components/Stack';
 import DetailsCategory from './DetailsCategory';
-import { mockReportData } from '../../data/reportData';
+import { mockReportData } from '../../common/reportData';
 
 const Subtitle = styled.p`
   text-transform: uppercase;
