--- packages/client/src/examples/server-side-rendering/ssr.tsx
@@ -1,28 +1,0 @@
-/* Example of service rendering in case if we need to build any crazy analytics or improve SEO */
-import React from 'react';
-import { useQuery } from '@apollo/react-hooks';
-import gql from 'graphql-tag';
-import { withApollo } from '../../apollo/apollo';
-
-const QUERY = gql`
-  query GetHello {
-    hello
-  }
-`;
-
-const SSR = () => {
-  const { data, loading, error, refetch } = useQuery(QUERY);
-
-  if (loading) return <p>Loading...</p>;
-  if (error) return <p>Error: {error.message}</p>;
-
-  return (
-    <div>
-      <h1>This should be rendered on server side</h1>
-      <pre>Data: {data.hello}</pre>
-      <button onClick={() => refetch()}>Refetch</button>
-    </div>
-  );
-};
-
-export default withApollo({ ssr: true })(SSR);
